Refactor component imports to match new file structure
